730 fully separate os metrics
